Log-in/Log-out functionality added.
