+++ app/src/main/java/com/example/HSUAppProject/SignUp.java
@@ -36,7 +36,7 @@ public class SignUp extends AppCompatActivity {
 
         sp = getSharedPreferences("UserPref", Context.MODE_PRIVATE);
         if(sp.getString("uID", null) != null){
-            startActivity(new Intent(getApplicationContext(), MainActivity.class));
+            startActivity(new Intent(getApplicationContext(), HomeScreen.class));
         }
 
         regID = findViewById(R.id.signUpID);
@@ -100,7 +100,7 @@ public class SignUp extends AppCompatActivity {
                 edit.commit();
 
                 Toast.makeText(SignUp.this, "Registered Successfully!", Toast.LENGTH_SHORT).show();
-                startActivity(new Intent(getApplicationContext(), MainActivity.class));
+                startActivity(new Intent(getApplicationContext(), HomeScreen.class));
             }
         });
