+++ app/src/main/java/com/example/HSUAppProject/ui/UserSettings.java
@@ -1,18 +1,34 @@
 package com.example.HSUAppProject.ui;
 
+import android.content.Context;
+import android.content.Intent;
+import android.content.SharedPreferences;
 import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+import android.widget.Toast;
 
 import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.preference.PreferenceFragmentCompat;
 
+import com.example.HSUAppProject.Login;
 import com.example.HSUAppProject.R;
 
 public class UserSettings extends AppCompatActivity {
+    Button mainBtn, appCreateTest;
+    SharedPreferences sp;
+    SharedPreferences.Editor edit;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+
+        sp = getSharedPreferences("UserPref", Context.MODE_PRIVATE);
+        if(sp.getString("uID", null) == null){
+            startActivity(new Intent(getApplicationContext(), Login.class));
+        }
+
         setContentView(R.layout.settings_activity);
         if (savedInstanceState == null) {
             getSupportFragmentManager()
@@ -24,12 +40,30 @@ public class UserSettings extends AppCompatActivity {
         if (actionBar != null) {
             actionBar.setDisplayHomeAsUpEnabled(true);
         }
+
+        mainBtn = findViewById(R.id.mainLogOut);
+        mainBtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                edit = sp.edit();
+                edit.putString("uID", null);
+                edit.commit();
+
+                Toast.makeText(UserSettings.this, "Logged Out Successfully!", Toast.LENGTH_SHORT).show();
+                startActivity(new Intent(getApplicationContext(), Login.class));
+            }
+        });
+
+
     }
+
 
     public static class SettingsFragment extends PreferenceFragmentCompat {
         @Override
         public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
             setPreferencesFromResource(R.xml.root_preferences, rootKey);
         }
+
+
     }
 }
