+++ app/src/main/res/layout/settings_activity.xml
@@ -5,5 +5,12 @@
     <FrameLayout
         android:id="@+id/settings"
         android:layout_width="match_parent"
-        android:layout_height="match_parent" />
+        android:layout_height="match_parent" >
+
+        <Button
+            android:id="@+id/mainLogOut"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/mainLogOut" />
+    </FrameLayout>
 </LinearLayout>
